+++ code/services/tasks/src/test/resources/init.sql
@@ -11,7 +11,7 @@ CREATE TABLE task (
     resolved INTEGER NOT NULL,
     rejected INTEGER NOT NULL,
     note VARCHAR(250) NULL,
-    last_updated INTEGER NOT NULL);
+    last_updated BIGINT NOT NULL);
 
 -- ------------------------------------------------------------
 -- Add Some dummy data
